Made the Order fully immutable. Made the Disciple run really parallel. Added timeout and loop exit condition for the Sensei(prepare) and Delivery(deliver) operations. Added the ThreadsSequentialAndParallelTest. Updated the UML

+++ src/main/java/org/sweetieslab/service/ManagementService.java
@@ -1,6 +1,5 @@
 package org.sweetieslab.service;
 
-import java.util.EnumMap;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -8,8 +7,6 @@ import java.util.concurrent.locks.ReentrantLock;
 import java.util.logging.Logger;
 import org.sweetieslab.model.order.Address;
 import org.sweetieslab.model.order.Order;
-import org.sweetieslab.model.pancakes.Ingredient;
-import org.sweetieslab.model.pancakes.PancakeFactory;
 import org.sweetieslab.model.pancakes.PancakeRecipe;
 
 public class ManagementService {
@@ -50,9 +47,18 @@ public class ManagementService {
     }
 
     public void addPancakes(PancakeRecipe pancake, UUID orderId, int count) {
-        Order order = dataService.addPancakes(pancake, orderId, count);
-        LOGGER.info(ADDED_PANCAKES_MESSAGE.formatted(count, pancake, order.getId(),
-            order.getPancakesCount(), order.getBuilding(), order.getRoom()));
+        dataService.addPancakes(pancake, orderId, count);
+        Order order = dataService.getOrder(orderId);
+        LOGGER.info(ADDED_PANCAKES_MESSAGE.formatted(count, pancake, orderId, count,
+            order.getBuilding(), order.getRoom()));
+    }
+
+    public void removePancakes(PancakeRecipe pancake, UUID orderId, int count) {
+        dataService.removePancakes(pancake, orderId, count);
+        Order order = dataService.getOrder(orderId);
+        int pancakesCount = dataService.getPancakesCount(orderId);
+        LOGGER.info(REMOVED_PANCAKES_MESSAGE.formatted(count, pancake, order.getId(),
+            pancakesCount, order.getBuilding(), order.getRoom()));
     }
 
     public Order getOrder(UUID orderId) {
@@ -60,42 +66,41 @@ public class ManagementService {
     }
 
     public List<String> viewOrder(UUID orderId) {
-        return dataService.getOrder(orderId).getPancakesDescriptions();
+        return dataService.getPancakesDescriptions(orderId);
     }
 
-    public void removePancakes(EnumMap<Ingredient, Integer> ingredients, UUID orderId, int count) {
-        removePancakes(PancakeFactory.getPancakeRecipe(ingredients), orderId, count);
-    }
-
-    public void removePancakes(PancakeRecipe pancake, UUID orderId, int count) {
-        Order order = dataService.removePancakes(pancake, orderId, count);
-        LOGGER.info(REMOVED_PANCAKES_MESSAGE.formatted(count, pancake, order.getId(),
-            order.getPancakesCount(), order.getBuilding(), order.getRoom()));
-    }
-
-    public void cancelOrder(UUID orderId) {
-        Order order = dataService.removeOrder(orderId);
-        LOGGER.info(CANCELLED_ORDER_MESSAGE.formatted(order.getId(), order.getPancakesCount(),
+    public void cancelOrder(Order order) {
+        int pancakesCount = dataService.getPancakesCount(order.getId());
+        dataService.removeOrder(order.getId());
+        LOGGER.info(CANCELLED_ORDER_MESSAGE.formatted(order.getId(), pancakesCount,
             order.getBuilding(), order.getRoom()));
     }
 
-    public synchronized void completeOrder(UUID orderId) {
-        Order order = dataService.completeOrder(orderId);
-        operationsService.completeOrder(order);
-        LOGGER.info(COMPLETED_ORDER_MESSAGE.formatted(order.getId(), order.getPancakesCount(),
-            order.getBuilding(), order.getRoom()));
+    public void completeOrder(Order order) {
+        if (dataService.orderIsEmpty(order.getId())) {
+            cancelOrder(order);
+        } else {
+            operationsService.completeOrder(order);
+            int pancakesCount = dataService.getPancakesCount(order.getId());
+            LOGGER.info(COMPLETED_ORDER_MESSAGE.formatted(order.getId(), pancakesCount,
+                order.getBuilding(), order.getRoom()));
+        }
     }
 
     public Set<UUID> listCompletedOrders() {
-        return dataService.listCompletedOrders();
+        return operationsService.listCompletedOrders();
     }
 
     public Order prepareOrder() {
         Order order = operationsService.prepareOrder();
-        dataService.prepareOrder(order.getId());
+        if (order == null) {
+            return null;
+        }
+        UUID orderId = order.getId();
+        int pancakesCount = dataService.getPancakesCount(orderId);
         prepareDeliverLoggingLock.lock();
         try {
-            LOGGER.info(PREPARED_ORDER_MESSAGE.formatted(order.getId(), order.getPancakesCount(),
+            LOGGER.info(PREPARED_ORDER_MESSAGE.formatted(orderId, pancakesCount,
                 order.getBuilding(), order.getRoom()));
         } finally {
             prepareDeliverLoggingLock.unlock();
@@ -104,16 +109,20 @@ public class ManagementService {
     }
 
     public Set<UUID> listPreparedOrders() {
-        return dataService.listPreparedOrders();
+        return operationsService.listPreparedOrders();
     }
 
     public Order deliverOrder() {
         Order order = operationsService.deliverOrder();
+        if (order == null) {
+            return null;
+        }
         UUID orderId = order.getId();
+        int pancakesCount = dataService.getPancakesCount(orderId);
         dataService.removeOrder(orderId);
         prepareDeliverLoggingLock.lock();
         try {
-            LOGGER.info(DELIVERED_ORDER_MESSAGE.formatted(order.getId(), order.getPancakesCount(),
+            LOGGER.info(DELIVERED_ORDER_MESSAGE.formatted(order.getId(), pancakesCount,
                 order.getBuilding(), order.getRoom()));
         } finally {
             prepareDeliverLoggingLock.unlock();
